First test not finished yet

+++ cdm/src/main/java/ucar/nc2/ft2/coverage/simpgeo/CFPoint.java
@@ -84,7 +84,7 @@ public class CFPoint {
 	 * @param next - next point if part of a multipoint
 	 * @param prev - previous point if part of a multipoint
 	 */
-	public CFPoint(int x, int y, CFPoint next, CFPoint prev)
+	public CFPoint(double x, double y, CFPoint next, CFPoint prev)
 	{
 		this.next = next;
 		this.prev = prev;
